refactor: packaged to kafka
