+++ src/main/java/ChatManagement/kafka/service/KafkaService.java
@@ -1,6 +1,6 @@
-package ChatManagement.service;
+package ChatManagement.kafka.service;
 
-import ChatManagement.domain.ChatMessage;
+import ChatManagement.kafka.domain.ChatMessage;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
